pyproject.toml
[project]
name = "cpc-commentary"
version = "0.1.0"
description = "Presentation commentary bots powered by Claude Code"
requires-python = ">=3.12"
dependencies = [
    "slack-bolt>=1.27.0",
    "slack-sdk>=3.41.0",
    "python-dotenv>=1.0.1",
    "aiohttp>=3.11.13",
    "pymupdf>=1.24.0",
    "sounddevice>=0.5.0",
    "faster-whisper>=1.0.0",
    "numpy>=1.26.0",
    "pyobjc-framework-quartz>=12.1",
]
